searchapi/pyproject.toml
[project]
name = "searchapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "phillip-chae", email = "msc694@nyu.edu" }
]
requires-python = ">=3.13"
dependencies = [
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "shared",
    "pillow-heif>=1.1.1",
    "imagehash>=4.3.2",
    "dependency-injector>=4.48.2",
    "uuid7>=0.1.0",
    "pymilvus (>=2.6.0,<3.0.0)",
    "uvicorn>=0.38.0",
    "python-magic>=0.4.27",
    "fastapi>=0.121.2",
    "transformers>=4.57.3",
    "open-clip-torch>=3.2.0",
]

[project.optional-dependencies]
gpu = ["shared[torch-gpu]"]
cpu = ["shared[torch-cpu]"]

[tool.uv]
package = false

[tool.uv.sources]
shared = { workspace = true }